Modularize & deprecate memory config
