+++ openfisca_core/memory_config.py
@@ -1,21 +1,9 @@
-# -*- coding: utf-8 -*-
+# The memory config module has been deprecated since X.X.X,
+# and will be removed in the future.
+#
+# Module's contents have been moved to the experimental module.
+#
+# The following are transitional imports to ensure non-breaking changes.
+# Could be deprecated in the next major release.
 
-import logging
-
-log = logging.getLogger(__name__)
-
-
-class MemoryConfig(object):
-
-    def __init__(self,
-      max_memory_occupation,
-      priority_variables = None,
-      variables_to_drop = None):
-        log.warn("Memory configuration is a feature that is still currently under experimentation. You are very welcome to use it and send us precious feedback, but keep in mind that the way it is used might change without any major version bump.")
-
-        self.max_memory_occupation = float(max_memory_occupation)
-        if self.max_memory_occupation > 1:
-            raise ValueError("max_memory_occupation must be <= 1")
-        self.max_memory_occupation_pc = self.max_memory_occupation * 100
-        self.priority_variables = set(priority_variables) if priority_variables else set()
-        self.variables_to_drop = set(variables_to_drop) if variables_to_drop else set()
+from openfisca_core.experimental import MemoryConfig  # noqa: F401
